refactored handling of IntegrationPoint, removing android specific concerns to Android-only code
